Updated decoding

+++ worker_R128.py
@@ -16,65 +16,206 @@ from ioMicro import *
 ncols = 4
 
 psf_file = r"psf_D103_B.npy"
-save_folder = r"/mnt/merfish9/20230919_R128_N5S1MER_analysis/RMERFISH/"
-flat_field_fl = r"lemon__med_col_raw"
+#save_folder = r"/mnt/merfish11/20230919_R128_N5S1MERNeuro_analysis/RMERFISH/"
+#save_folder = r"/mnt/merfish11/20230919_R128_N5S1MERControl_analysis/RMERFISH/"
+save_folder = r"/mnt/renimaging2/02_20_2024_BigHeartR133S2_analysis/"
+flat_field_fl = r"R128__med_col_raw"
+lib_fl = "/home/plt3/MERlin_Parameters/codebooks/codebook_code_color2__QZBB_P3P4Heart_blank.csv"
 
+def get_icodesV3(dec,nmin_bits=3,iH=-3):
+    import time
+    start = time.time()
+    lens = dec.lens
+    res_unfolder = dec.res_unfolder
+    Mlen = np.max(lens)
+    print("Calculating indexes within cluster...")
+    res_is = np.tile(np.arange(Mlen), len(lens))
+    res_is = res_is[res_is < np.repeat(lens, Mlen)]
+    print("Calculating index of molecule...")
+    ires = np.repeat(np.arange(len(lens)), lens)
+    #r0 = np.array([r[0] for r in res for r_ in r])
+    print("Calculating index of first molecule...")
+    r0i = np.concatenate([[0],np.cumsum(lens)])[:-1]
+    r0 = res_unfolder[np.repeat(r0i, lens)]
+    print("Total time unfolded molecules:",time.time()-start)
+    
+    ### torch
+    ires = torch.from_numpy(ires.astype(np.int64))
+    res_unfolder = torch.from_numpy(res_unfolder.astype(np.int64))
+    res_is = torch.from_numpy(res_is.astype(np.int64))
+    
+    import time
+    start = time.time()
+    print("Computing score...")
+    scoreF = torch.from_numpy(dec.XH[:,iH])[res_unfolder]
+    print("Total time computing score:",time.time()-start)
+    
+    
+    ### organize molecules in blocks for each cluster
+    def get_asort_scores():
+        val = torch.max(scoreF)+2
+        scoreClu = torch.zeros([len(lens),Mlen],dtype=torch.float64)+val
+        scoreClu[ires,res_is]=scoreF
+        asort = scoreClu.argsort(-1)
+        scoreClu = torch.gather(scoreClu,dim=-1,index=asort)
+        scoresF2 = scoreClu[scoreClu<val-1]
+        return asort,scoresF2
+    def get_reorder(x,val=-1):
+        if type(x) is not torch.Tensor:
+            x = torch.from_numpy(np.array(x))
+        xClu = torch.zeros([len(lens),Mlen],dtype=x.dtype)+val
+        xClu[ires,res_is] = x
+        xClu = torch.gather(xClu,dim=-1,index=asort)
+        xf = xClu[xClu>val]
+        return xf
+    
+    
+    import time
+    start = time.time()
+    print("Computing sorting...")
+    asort,scoresF2 = get_asort_scores()
+    res_unfolder2 = get_reorder(res_unfolder,val=-1)
+    del asort
+    del scoreF
+    print("Total time sorting molecules by score:",time.time()-start)
+    
+    
+    
+    import time
+    start = time.time()
+    print("Finding best bits per molecules...")
+    
+    Rs = dec.XH[:,-1].astype(np.int64)
+    Rs = torch.from_numpy(Rs)
+    Rs_U = Rs[res_unfolder2]
+    nregs,nbits = dec.codes_01.shape
+    score_bits = torch.zeros([len(lens),nbits],dtype=scoresF2.dtype)-1
+    score_bits[ires,Rs_U]=scoresF2
+    
+    
+    codes_lib = torch.from_numpy(np.array(dec.codes__))
+    
+    
+    codes_lib_01 = torch.zeros([len(codes_lib),nbits],dtype=score_bits.dtype)
+    for icd,cd in enumerate(codes_lib):
+        codes_lib_01[icd,cd]=1
+    codes_lib_01 = codes_lib_01/torch.norm(codes_lib_01,dim=-1)[:,np.newaxis]
+    print("Finding best code...")
+    batch = 10000
+    icodes_best = torch.zeros(len(score_bits),dtype=torch.int64)
+    dists_best = torch.zeros(len(score_bits),dtype=torch.float32)
+    from tqdm import tqdm
+    for i in tqdm(range((len(score_bits)//batch)+1)):
+        score_bits_ = score_bits[i*batch:(i+1)*batch]
+        if len(score_bits_)>0:
+            score_bits__ = score_bits_.clone()
+            score_bits__[score_bits__==-1]=0
+            score_bits__ = score_bits__/torch.norm(score_bits__,dim=-1)[:,np.newaxis]
+            Mul = torch.matmul(score_bits__,codes_lib_01.T)
+            max_ = torch.max(Mul,dim=-1)
+            icodes_best[i*batch:(i+1)*batch] = max_.indices
+            dists_best[i*batch:(i+1)*batch] = 2-2*max_.values
+    
+    
+    keep_all_bits = torch.sum(score_bits.gather(1,codes_lib[icodes_best])>=0,-1)>=nmin_bits
+    dists_best_ = dists_best[keep_all_bits]
+    score_bits = score_bits[keep_all_bits]
+    icodes_best_ = icodes_best[keep_all_bits]
+    icodesN=icodes_best_
+    
+    indexMols_ = torch.zeros([len(lens),nbits],dtype=res_unfolder2.dtype)-1
+    indexMols_[ires,Rs_U]=res_unfolder2
+    indexMols_ = indexMols_[keep_all_bits]
+    indexMols_ = indexMols_.gather(1,codes_lib[icodes_best_])
+    
+    # make unique
+    indexMols_,rinvMols = get_unique_ordered(indexMols_)
+    icodesN = icodesN[rinvMols]
+    
+    XH = torch.from_numpy(dec.XH)
+    XH_pruned = XH[indexMols_]
+    XH_pruned[indexMols_==-1]=np.nan
+    
+    dec.dist_best = dists_best_[rinvMols].numpy()
+    dec.XH_pruned=XH_pruned.numpy()
+    dec.icodesN=icodesN.numpy()
+    np.savez_compressed(dec.decoded_fl,XH_pruned=dec.XH_pruned,icodesN=dec.icodesN,gns_names = np.array(dec.gns_names),dist_best=dec.dist_best)
+    print("Total time best bits per molecule:",time.time()-start)
 
-def compute_drift(save_folder, fov, all_flds, set_, redo=False):
-    """
-    save_folder where to save analyzed data
-    fov - i.e. Conv_zscan_005.zarr
-    all_flds - folders that contain eithger the MERFISH bits or control bits or smFISH
-    set_ - an extra tag typically at the end of the folder to separate out different folders
-    """
-    # print(len(all_flds))
-    # print(all_flds)
-    gpu = False
-    # defulat name of the drift file
-    drift_fl = save_folder + os.sep + "driftNew_" + fov.split(".")[0] + "--" + set_ + ".pkl"
+def compute_decoding(save_folder,fov,set_,redo=False):
+    dec = decoder_simple(save_folder,fov,set_)
+    #self.decoded_fl = self.decoded_fl.replace('decoded_','decodedNew_')
+    complete = dec.check_is_complete()
+    if complete==0 or redo:
+        #compute_drift(save_folder,fov,all_flds,set_,redo=False,gpu=False)
+        dec = decoder_simple(save_folder,fov=fov,set_=set_)
+        dec.get_XH(fov,set_,ncols=3,nbits=16,th_h=0)#number of colors match 
+        print(dec.XH.shape)
+        dec.XH = dec.XH[dec.XH[:,-4]>0.25] ### keep the spots that are correlated with the expected PSF for 60X
+        dec.load_library(lib_fl,nblanks=-1)
+        
+        dec.ncols = 3
+        #dec.get_inters(dinstance_th=2,enforce_color=True)# enforce_color=False
+        dec.get_inters(dinstance_th=2,nmin_bits=3,enforce_color=True,redo=True)
+        #dec.get_icodes(nmin_bits=4,method = 'top4',norm_brightness=None,nbits=24)#,is_unique=False)
+        get_icodesV3(dec,nmin_bits=3,iH=-3)
+        #get_icodesV3(dec,iH=-3,redo=False,norm_brightness=False,nbits=48,is_unique=True)
 
-    iiref = None
-    fl_ref = None
-    previous_drift = {}
-    if not os.path.exists(drift_fl) or redo:
-        redo = True
+"""
+def get_best_translation_pointsV2(fl,fl_ref,save_folder,set_,resc=5):
+    
+    obj = get_dapi_features(fl,save_folder,set_)
+    obj_ref = get_dapi_features(fl_ref,save_folder,set_)
+    tzxyf,tzxy_plus,tzxy_minus,N_plus,N_minus = np.array([0,0,0]),np.array([0,0,0]),np.array([0,0,0]),0,0
+    if (len(obj.Xh_plus)>0) and (len(obj.Xh_minus)>0) and  (len(obj_ref.Xh_plus)>0) and (len(obj_ref.Xh_minus)>0):
+        X = obj.Xh_plus[:,:3]
+        X_ref = obj_ref.Xh_plus[:,:3]
+        tzxy_plus,N_plus = get_best_translation_points(X,X_ref,resc=resc,return_counts=True)
+
+        X = obj.Xh_minus[:,:3]
+        X_ref = obj_ref.Xh_minus[:,:3]
+        tzxy_minus,N_minus = get_best_translation_points(X,X_ref,resc=resc,return_counts=True)
+        
+        tzxyf = -(tzxy_plus+tzxy_minus)/2
+        
+    
+
+    return [tzxyf,tzxy_plus,tzxy_minus,N_plus,N_minus]
+"""
+
+def get_best_translation_pointsV2(fl,fl_ref,save_folder,set_,resc=5):
+    
+    obj = get_dapi_features(fl,save_folder,set_)
+    obj_ref = get_dapi_features(fl_ref,save_folder,set_)
+    tzxyf,tzxy_plus,tzxy_minus,N_plus,N_minus = np.array([0,0,0]),np.array([0,0,0]),np.array([0,0,0]),0,0
+    if (len(obj.Xh_plus)>0) and (len(obj_ref.Xh_plus)>0):
+        X = obj.Xh_plus[:,:3]
+        X_ref = obj_ref.Xh_plus[:,:3]
+        tzxy_plus,N_plus = get_best_translation_points(X,X_ref,resc=resc,return_counts=True)
+    if (len(obj.Xh_minus)>0) and (len(obj_ref.Xh_minus)>0):
+        X = obj.Xh_minus[:,:3]
+        X_ref = obj_ref.Xh_minus[:,:3]
+        tzxy_minus,N_minus = get_best_translation_points(X,X_ref,resc=resc,return_counts=True)
+    if np.max(np.abs(tzxy_minus-tzxy_plus))<=2:
+        tzxyf = -(tzxy_plus*N_plus+tzxy_minus*N_minus)/(N_plus+N_minus)
     else:
-        try:
-            drifts_, all_flds_, fov_, fl_ref = pickle.load(open(drift_fl, "rb"))
-            all_tags_ = np.array([os.path.basename(fld) for fld in all_flds_])
-            all_tags = np.array([os.path.basename(fld) for fld in all_flds])
-            iiref = np.argmin([np.sum(np.abs(drift[0])) for drift in drifts_])
-            previous_drift = {tag: drift for drift, tag in zip(drifts_, all_tags_)}
+        tzxyf = -[tzxy_plus,tzxy_minus][np.argmax([N_plus,N_minus])]
+    
 
-            if not (len(all_tags_) == len(all_tags)):
-                redo = True
-            else:
-                if not np.all(np.sort(all_tags_) == np.sort(all_tags)):
-                    redo = True
-        except:
-            os.remove(drift_fl)
-            redo = True
-    if redo:
-        fls = [fld + os.sep + fov for fld in all_flds]
-        if fl_ref is None:
-            fl_ref = fls[len(fls) // 2]
-        obj = None
+    return [tzxyf,tzxy_plus,tzxy_minus,N_plus,N_minus]
+
+
+def compute_drift_V2(save_folder,fov,all_flds,set_,redo=False,gpu=False):
+    drift_fl = save_folder+os.sep+'driftNew_'+fov.split('.')[0]+'--'+set_+'.pkl'
+    if not os.path.exists(drift_fl) or redo:
+        fls = [fld+os.sep+fov for fld in all_flds]
+        fl_ref = fls[len(fls)//2]
         newdrifts = []
-        all_fldsT = []
-        for fl in tqdm(fls):
-            fld = os.path.dirname(fl)
-            tag = os.path.basename(fld)
-            new_drift_info = previous_drift.get(tag, None)
-            if new_drift_info is None:
-                if obj is None:
-                    obj = fine_drift(fl_ref, fl, sz_block=600)
-                else:
-                    obj.get_drift(fl_ref, fl)
-                new_drift = -(obj.drft_minus + obj.drft_plus) / 2
-                new_drift_info = [new_drift, obj.drft_minus, obj.drft_plus, obj.drift, obj.pair_minus, obj.pair_plus]
-            newdrifts.append(new_drift_info)
-            all_fldsT.append(fld)
-            pickle.dump([newdrifts, all_fldsT, fov, fl_ref], open(drift_fl, "wb"))
+        for fl in fls:
+            drft = get_best_translation_pointsV2(fl,fl_ref,save_folder,set_,resc=5)
+            print(drft)
+            newdrifts.append(drft)
+        pickle.dump([newdrifts,all_flds,fov,fl_ref],open(drift_fl,'wb'))
 
 
 def compute_drift_features(save_folder, fov, all_flds, set_, redo=True, gpu=True):
@@ -159,38 +300,30 @@ def compute_fits(
                     main_do_compute_fits(save_folder, fld, fov, icol, save_fl, psf, old_method, redefine_color, gpu)
 
 
-def compute_decoding(save_folder, fov, set_):
-    dec = decoder_simple(save_folder, fov, set_)
-    complete = dec.check_is_complete()
-    if complete == 0:
-        dec.get_XH(fov, set_, ncols=ncols)  # number of colors match
-        dec.XH = dec.XH[dec.XH[:, -4] > 0.25]  ### keep the spots that are correlated with the expected PSF for 60X
-        dec.load_library(
-            lib_fl=r"\\192.168.0.10\bbfishdc13\codebook_0_New_DCBB-300_MERFISH_encoding_2_21_2023.csv", nblanks=-1
-        )
-        dec.get_inters(dinstance_th=2, enforce_color=True)  # enforce_color=False
-        dec.get_icodes(nmin_bits=4, method="top4", norm_brightness=-1)
-
 
 def main_f(fov, try_mode=True, gpu=True):
-    #print("Computing fitting on: " + str(fov))
-    compute_fits(save_folder, fov, all_flds, redo=False, try_mode=try_mode, redefine_color=redefine_color, gpu=gpu)
+    print("Computing fitting on: " + str(fov))
+    #compute_fits(save_folder, fov, all_flds, redo=False, try_mode=try_mode, redefine_color=redefine_color, gpu=gpu)
     #print("Computing drift on: " + str(fov))
     #try:
-    #    compute_drift_features(save_folder, fov, all_flds, "", redo=False, gpu=False)
+        #compute_drift_features(save_folder, fov, all_flds, "", redo=False, gpu=False)
+    compute_drift_V2(save_folder,fov,all_flds,"",redo=False)
+    compute_decoding(save_folder,fov,"",redo=False)
     #except Exception as e:
     #    print("corruption at: " + fov, e)
-    # compute_drift(save_folder,fov,all_flds,set_,redo=False)
-
-    # compute_decoding(save_folder,fov,set_)
-
-
+    
+def cleanup_decoded_drift(save_folder):
+    all_fls = glob.glob(save_folder+os.sep+'decoded*')
+    all_fls += glob.glob(save_folder+os.sep+'drift*')
+    for fl in tqdm(all_fls):
+        os.remove(fl)
 if __name__ == "__main__":
     if not os.path.exists(save_folder):
         os.makedirs(save_folder)
-    set__ = ""
+    set__ = "set2"
 
-    all_flds = glob.glob(r"/mnt/merfish9/20230919_R128_N5S1MER/RMERFISH/H*RMER*" + set__ + os.sep) 
+    all_flds = glob.glob(r"/mnt/renimaging2/02_20_2024_BigHeartR133S2/H*" + set__ + os.sep)
+    #all_flds += glob.glob(r"/mnt/merfish12/20231024_RD129_N5S2heart/RMERFISH/H*RMER*" + set__ + os.sep)
     redefine_color = [[0, 1, 2, 3]] * len(all_flds)  #### three signal color
 
     # print(all_flds)
@@ -203,34 +336,48 @@ if __name__ == "__main__":
     else:
         fovs = np.load(fovs_fl)
 
-    from queue import Empty
+    fovs=np.sort(fovs)
+    #cleanup_decoded_drift(save_folder)
+    #main_f(fovs[5])
+    if True:
 
-    def worker(queue, gpu):
+        from queue import Empty
+
+        def worker(queue, gpu):
+            try:
+                while True:
+                    try:
+                        fov = queue.get(False)
+                        main_f(fov, gpu=gpu)
+                    except FileNotFoundError as e:
+                        print(fov, e)
+            except Empty:
+                return
+
+        import multiprocessing
+
+        queue = multiprocessing.Queue()
+        for item in sorted(fovs):
+            queue.put(item)
+
+        workers = [ ### Add however many GPU and CPU workers as you want here
+            multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": True}),
+            multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
+            multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
+            multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
+            multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
+            multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
+            multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
+            multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
+            multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
+        ]
         try:
-            while True:
-                main_f(queue.get(False), gpu=gpu)
-        except Empty:
-            return
-
-    import multiprocessing
-
-    queue = multiprocessing.Queue()
-    for item in sorted(fovs):
-        queue.put(item)
-
-    workers = [ ### Add however many GPU and CPU workers as you want here
-        multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": True}),
-        multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
-        multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
-        #multiprocessing.Process(target=worker, kwargs={"queue": queue, "gpu": False}),
-    ]
-    try:
-        for worker in workers:
-            worker.start()
-        for worker in workers:
-            worker.join()
-            worker.close()
-    except KeyboardInterrupt:
-        for worker in workers:
-            worker.terminate()
-            worker.close()
+            for worker in workers:
+                worker.start()
+            for worker in workers:
+                worker.join()
+                worker.close()
+        except KeyboardInterrupt:
+            for worker in workers:
+                worker.terminate()
+                worker.close()
